SubGraph : Adopt reference loading features from Reference
The long term plan here is to remove Reference entirely, with SubGraph being able to support both referencing and local modes.
